+++ src/extension/notebook/renderer/json_entry.tsx
@@ -21,43 +21,29 @@
  * SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
 
-import * as React from 'react';
-import {Result, ResultJSON} from '@malloydata/malloy';
-import {HTMLView} from '@malloydata/render';
+import ReactDOM from 'react-dom';
+import React from 'react';
+import {StyleSheetManager} from 'styled-components';
+import {ActivationFunction} from 'vscode-notebook-renderer';
+import {JsonRenderer} from './JsonRenderer';
 
-export interface ResultProps {
-  results: ResultJSON;
-}
-
-export const MalloyRenderer: React.FC<ResultProps> = ({results}) => {
-  const [html, setHtml] = React.useState<HTMLElement>();
-
-  React.useEffect(() => {
-    if (results) {
-      const {data} = Result.fromJSON(results);
-      const rendering = new HTMLView(document).render(data, {dataStyles: {}});
-      rendering.then(rendered => {
-        setHtml(rendered);
-      });
-    }
-  }, [results]);
-
-  if (html) {
-    return <DOMElement element={html} />;
-  }
-  return null;
-};
-
-const DOMElement: React.FC<{element: HTMLElement}> = ({element}) => {
-  const ref = React.useRef<HTMLDivElement>(null);
-
-  React.useEffect(() => {
-    const parent = ref.current;
-    if (parent) {
-      parent.innerHTML = '';
-      parent.appendChild(element);
-    }
-  }, [element]);
-
-  return <div ref={ref} />;
+export const activate: ActivationFunction = () => {
+  return {
+    renderOutputItem(info, element) {
+      let shadow = element.shadowRoot;
+      if (!shadow) {
+        shadow = element.attachShadow({mode: 'open'});
+        const root = document.createElement('div');
+        root.id = 'root';
+        shadow.append(root);
+      }
+      const root = shadow.querySelector<HTMLElement>('#root');
+      ReactDOM.render(
+        <StyleSheetManager target={root}>
+          <JsonRenderer results={info.json()} />
+        </StyleSheetManager>,
+        root
+      );
+    },
+  };
 };
